fastapi-raw: move helper inside app module
